Move new methods to CompareUtilities + update tests
