+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/test/utils/TestingUtilities.java
@@ -71,7 +71,6 @@ import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSyntaxException;
 
 public class TestingUtilities extends BaseTestingUtilities {
-  private static final boolean SHOW_DIFF = true;
 
   static public Map<String, IWorkerContext> fcontexts;
 
@@ -178,352 +177,5 @@ public class TestingUtilities extends BaseTestingUtilities {
     throw new Error("FHIR US directory not configured");
   }
 
-  public static String createNotEqualMessage(final String message, final String expected, final String actual) {
-    return new StringBuilder()
-      .append(message).append('\n')
-      .append("Expected :").append(expected).append('\n')
-      .append("Actual  :").append(actual).toString();
-  }
 
-  public static String checkXMLIsSame(InputStream expected, InputStream actual) throws Exception {
-    String result = compareXml(expected, actual);
-    return result;
-  }
-
-  public static String checkXMLIsSame(String expected, String actual) throws Exception {
-    String result = compareXml(expected, actual);
-    if (result != null && SHOW_DIFF) {
-      String diff = ToolGlobalSettings.hasComparePath() ? ToolGlobalSettings.getComparePath() : Utilities.path(System.getenv("ProgramFiles"), "WinMerge", "WinMergeU.exe");
-      if (new File(diff).exists() || Utilities.isToken(diff)) {
-        Runtime.getRuntime().exec(new String[]{diff, expected, actual});
-      }
-    }
-    return result;
-  }
-
-  private static String compareXml(InputStream actual, InputStream expected) throws Exception {
-    return compareElements("", loadXml(expected).getDocumentElement(), loadXml(actual).getDocumentElement());
-  }
-
-  private static String compareXml(String actual, String expected) throws Exception {
-    return compareElements("", loadXml(expected).getDocumentElement(), loadXml(actual).getDocumentElement());
-  }
-
-  private static String compareElements(String path, Element expectedElement, Element actualElement) {
-    if (!namespacesMatch(expectedElement.getNamespaceURI(), actualElement.getNamespaceURI()))
-      return createNotEqualMessage("Namespaces differ at " + path, expectedElement.getNamespaceURI(), actualElement.getNamespaceURI());
-    if (!expectedElement.getLocalName().equals(actualElement.getLocalName()))
-      return createNotEqualMessage("Names differ at " + path ,  expectedElement.getLocalName(), actualElement.getLocalName());
-    path = path + "/" + expectedElement.getLocalName();
-    String s = compareAttributes(path, expectedElement.getAttributes(), actualElement.getAttributes());
-    if (!Utilities.noString(s))
-      return s;
-    s = compareAttributes(path, expectedElement.getAttributes(), actualElement.getAttributes());
-    if (!Utilities.noString(s))
-      return s;
-
-    Node expectedChild = expectedElement.getFirstChild();
-    Node actualChild = actualElement.getFirstChild();
-    expectedChild = skipBlankText(expectedChild);
-    actualChild = skipBlankText(actualChild);
-    while (expectedChild != null && actualChild != null) {
-      if (expectedChild.getNodeType() != actualChild.getNodeType())
-        return createNotEqualMessage("node type mismatch in children of " + path, Integer.toString(expectedElement.getNodeType()), Integer.toString(actualElement.getNodeType()));
-      if (expectedChild.getNodeType() == Node.TEXT_NODE) {
-        if (!normalise(expectedChild.getTextContent()).equals(normalise(actualChild.getTextContent())))
-          return createNotEqualMessage("Text differs at " + path, normalise(expectedChild.getTextContent()).toString(), normalise(actualChild.getTextContent()).toString());
-      } else if (expectedChild.getNodeType() == Node.ELEMENT_NODE) {
-        s = compareElements(path, (Element) expectedChild, (Element) actualChild);
-        if (!Utilities.noString(s))
-          return s;
-      }
-
-      expectedChild = skipBlankText(expectedChild.getNextSibling());
-      actualChild = skipBlankText(actualChild.getNextSibling());
-    }
-    if (expectedChild != null)
-      return "node mismatch - more nodes in expected in children of " + path;
-    if (actualChild != null)
-      return "node mismatch - more nodes in actual in children of " + path;
-    return null;
-  }
-
-  private static boolean namespacesMatch(String ns1, String ns2) {
-    return ns1 == null ? ns2 == null : ns1.equals(ns2);
-  }
-
-  private static Object normalise(String text) {
-    String result = text.trim().replace('\r', ' ').replace('\n', ' ').replace('\t', ' ');
-    while (result.contains("  "))
-      result = result.replace("  ", " ");
-    return result;
-  }
-
-  private static String compareAttributes(String path, NamedNodeMap expected, NamedNodeMap actual) {
-    for (int i = 0; i < actual.getLength(); i++) {
-
-      Node actualNode = actual.item(i);
-      String actualNodeName = actualNode.getNodeName();
-      if (!(actualNodeName.equals("xmlns") || actualNodeName.startsWith("xmlns:"))) {
-        Node expectedNode = expected.getNamedItem(actualNodeName);
-        if (expectedNode == null)
-          return "Attributes differ at " + path + ": missing attribute " + actualNodeName;
-        if (!normalise(actualNode.getTextContent()).equals(normalise(expectedNode.getTextContent()))) {
-          byte[] actualBytes = unBase64(actualNode.getTextContent());
-          byte[] expectedBytes = unBase64(expectedNode.getTextContent());
-          if (!sameBytes(expectedBytes, actualBytes))
-            return createNotEqualMessage("Attributes differ at " + path, normalise(expectedNode.getTextContent()).toString(), normalise(actualNode.getTextContent()).toString()) ;
-        }
-      }
-    }
-    return null;
-  }
-
-  private static boolean sameBytes(byte[] b1, byte[] b2) {
-    if (b1.length == 0 || b2.length == 0)
-      return false;
-    if (b1.length != b2.length)
-      return false;
-    for (int i = 0; i < b1.length; i++)
-      if (b1[i] != b2[i])
-        return false;
-    return true;
-  }
-
-  private static byte[] unBase64(String text) {
-    return Base64.decodeBase64(text);
-  }
-
-  private static Node skipBlankText(Node node) {
-    while (node != null && (((node.getNodeType() == Node.TEXT_NODE) && Utilities.isWhitespace(node.getTextContent())) || (node.getNodeType() == Node.COMMENT_NODE)))
-      node = node.getNextSibling();
-    return node;
-  }
-
-  private static Document loadXml(String fn) throws Exception {
-    return loadXml(new FileInputStream(fn));
-  }
-
-  private static Document loadXml(InputStream fn) throws Exception {
-    DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-    factory.setFeature("http://apache.org/xml/features/disallow-doctype-decl", true);
-    factory.setFeature("http://xml.org/sax/features/external-general-entities", false);
-    factory.setFeature("http://xml.org/sax/features/external-parameter-entities", false);
-    factory.setFeature("http://apache.org/xml/features/nonvalidating/load-external-dtd", false);
-    factory.setXIncludeAware(false);
-    factory.setExpandEntityReferences(false);
-
-    factory.setNamespaceAware(true);
-    DocumentBuilder builder = factory.newDocumentBuilder();
-    return builder.parse(fn);
-  }
-
-  public static String checkJsonSrcIsSame(String expected, String actual) throws JsonSyntaxException, FileNotFoundException, IOException {
-    return checkJsonSrcIsSame(expected, actual, true);
-  }
-
-  public static String checkJsonSrcIsSame(String expectedString, String actualString, boolean showDiff) throws JsonSyntaxException, FileNotFoundException, IOException {
-    String result = compareJsonSrc(expectedString, actualString);
-    if (result != null && SHOW_DIFF && showDiff) {
-      String diff = null;
-      if (System.getProperty("os.name").contains("Linux"))
-        diff = Utilities.path("/", "usr", "bin", "meld");
-      else {
-        if (Utilities.checkFile("WinMerge", Utilities.path(System.getenv("ProgramFiles(X86)"), "WinMerge"), "\\WinMergeU.exe", null))
-          diff = Utilities.path(System.getenv("ProgramFiles(X86)"), "WinMerge", "WinMergeU.exe");
-        else if (Utilities.checkFile("WinMerge", Utilities.path(System.getenv("ProgramFiles(X86)"), "Meld"), "\\Meld.exe", null))
-          diff = Utilities.path(System.getenv("ProgramFiles(X86)"), "Meld", "Meld.exe");
-      }
-      if (diff == null || diff.isEmpty())
-        return result;
-
-      List<String> command = new ArrayList<String>();
-      String expected = Utilities.path("[tmp]", "expected" + expectedString.hashCode() + ".json");
-      String actual = Utilities.path("[tmp]", "actual" + actualString.hashCode() + ".json");
-      TextFile.stringToFile(expectedString, expected);
-      TextFile.stringToFile(actualString, actual);
-      command.add(diff);
-      if (diff.toLowerCase().contains("meld"))
-        command.add("--newtab");
-      command.add(expected);
-      command.add(actual);
-
-      ProcessBuilder builder = new ProcessBuilder(command);
-      builder.directory(new CSFile(Utilities.path("[tmp]")));
-      builder.start();
-
-    }
-    return result;
-  }
-
-  public static String checkJsonIsSame(String expected, String actual) throws JsonSyntaxException, FileNotFoundException, IOException {
-    String result = compareJson(expected, actual);
-    if (result != null && SHOW_DIFF) {
-      String diff = Utilities.path(System.getenv("ProgramFiles(X86)"), "WinMerge", "WinMergeU.exe");
-      List<String> command = new ArrayList<String>();
-      command.add("\"" + diff + "\" \"" + expected +  "\" \"" + actual + "\"");
-
-      ProcessBuilder builder = new ProcessBuilder(command);
-      builder.directory(new CSFile("c:\\temp"));
-      builder.start();
-
-    }
-    return result;
-  }
-
-  private static String compareJsonSrc(String expected, String actual) throws JsonSyntaxException, FileNotFoundException, IOException {
-    JsonObject actualJsonObject = (JsonObject) new com.google.gson.JsonParser().parse(actual);
-    JsonObject expectedJsonObject = (JsonObject) new com.google.gson.JsonParser().parse(expected);
-    return compareObjects("", expectedJsonObject, actualJsonObject);
-  }
-
-  private static String compareJson(String expected, String actual) throws JsonSyntaxException, FileNotFoundException, IOException {
-    JsonObject actualJsonObject = (JsonObject) new com.google.gson.JsonParser().parse(TextFile.fileToString(actual));
-    JsonObject expectedJsonObject = (JsonObject) new com.google.gson.JsonParser().parse(TextFile.fileToString(expected));
-    return compareObjects("", expectedJsonObject, actualJsonObject);
-  }
-
-  private static String compareObjects(String path, JsonObject expectedJsonObject, JsonObject actualJsonObject) {
-    for (Map.Entry<String, JsonElement> en : actualJsonObject.entrySet()) {
-      String n = en.getKey();
-      if (!n.equals("fhir_comments")) {
-        if (expectedJsonObject.has(n)) {
-          String s = compareNodes(path + '.' + n, expectedJsonObject.get(n), en.getValue());
-          if (!Utilities.noString(s))
-            return s;
-        } else
-          return "properties differ at " + path + ": missing property " + n;
-      }
-    }
-    for (Map.Entry<String, JsonElement> en : expectedJsonObject.entrySet()) {
-      String n = en.getKey();
-      if (!n.equals("fhir_comments")) {
-        if (!actualJsonObject.has(n))
-          return "properties differ at " + path + ": missing property " + n;
-      }
-    }
-    return null;
-  }
-
-  private static String compareNodes(String path, JsonElement expectedJsonElement, JsonElement actualJsonElement) {
-    if (actualJsonElement.getClass() != expectedJsonElement.getClass())
-      return createNotEqualMessage("properties differ at " + path, expectedJsonElement.getClass().getName(), actualJsonElement.getClass().getName());
-    else if (actualJsonElement instanceof JsonPrimitive) {
-      JsonPrimitive actualJsonPrimitive = (JsonPrimitive) actualJsonElement;
-      JsonPrimitive expectedJsonPrimitive = (JsonPrimitive) expectedJsonElement;
-      if (actualJsonPrimitive.isBoolean() && expectedJsonPrimitive.isBoolean()) {
-        if (actualJsonPrimitive.getAsBoolean() != expectedJsonPrimitive.getAsBoolean())
-          return createNotEqualMessage("boolean property values differ at " + path , expectedJsonPrimitive.getAsString(), actualJsonPrimitive.getAsString());
-      } else if (actualJsonPrimitive.isString() && expectedJsonPrimitive.isString()) {
-        String actualJsonString = actualJsonPrimitive.getAsString();
-        String expectedJsonString = expectedJsonPrimitive.getAsString();
-        if (!(actualJsonString.contains("<div") && expectedJsonString.contains("<div")))
-          if (!actualJsonString.equals(expectedJsonString))
-            if (!sameBytes(unBase64(actualJsonString), unBase64(expectedJsonString)))
-              return createNotEqualMessage("string property values differ at " + path, expectedJsonString, actualJsonString);
-      } else if (actualJsonPrimitive.isNumber() && expectedJsonPrimitive.isNumber()) {
-        if (!actualJsonPrimitive.getAsString().equals(expectedJsonPrimitive.getAsString()))
-          return createNotEqualMessage("number property values differ at " + path, expectedJsonPrimitive.getAsString(), actualJsonPrimitive.getAsString());
-      } else
-        return createNotEqualMessage("property types differ at " + path, expectedJsonPrimitive.getAsString(), actualJsonPrimitive.getAsString());
-    } else if (actualJsonElement instanceof JsonObject) {
-      String s = compareObjects(path, (JsonObject) expectedJsonElement, (JsonObject) actualJsonElement);
-      if (!Utilities.noString(s))
-        return s;
-    } else if (actualJsonElement instanceof JsonArray) {
-      JsonArray actualArray = (JsonArray) actualJsonElement;
-      JsonArray expectedArray = (JsonArray) expectedJsonElement;
-
-      if (actualArray.size() != expectedArray.size())
-        return createNotEqualMessage("array properties count differs at " + path, Integer.toString(expectedArray.size()), Integer.toString(actualArray.size()));
-      for (int i = 0; i < actualArray.size(); i++) {
-        String s = compareNodes(path + "[" + Integer.toString(i) + "]", expectedArray.get(i), actualArray.get(i));
-        if (!Utilities.noString(s))
-          return s;
-      }
-    } else if (actualJsonElement instanceof JsonNull) {
-
-    } else
-      return "unhandled property " + actualJsonElement.getClass().getName();
-    return null;
-  }
-
-  public static String temp() {
-    if (new File("c:\\temp").exists())
-      return "c:\\temp";
-    return System.getProperty("java.io.tmpdir");
-  }
-
-  public static String checkTextIsSame(String expected, String actual) throws JsonSyntaxException, FileNotFoundException, IOException {
-    return checkTextIsSame(expected, actual, true);
-  }
-
-  public static String checkTextIsSame(String expectedString, String actualString, boolean showDiff) throws JsonSyntaxException, FileNotFoundException, IOException {
-    String result = compareText(expectedString, actualString);
-    if (result != null && SHOW_DIFF && showDiff) {
-      String diff = null;
-      if (System.getProperty("os.name").contains("Linux"))
-        diff = Utilities.path("/", "usr", "bin", "meld");
-      else {
-        if (Utilities.checkFile("WinMerge", Utilities.path(System.getenv("ProgramFiles(X86)"), "WinMerge"), "\\WinMergeU.exe", null))
-          diff = Utilities.path(System.getenv("ProgramFiles(X86)"), "WinMerge", "WinMergeU.exe");
-        else if (Utilities.checkFile("WinMerge", Utilities.path(System.getenv("ProgramFiles(X86)"), "Meld"), "\\Meld.exe", null))
-          diff = Utilities.path(System.getenv("ProgramFiles(X86)"), "Meld", "Meld.exe");
-      }
-      if (diff == null || diff.isEmpty())
-        return result;
-
-      List<String> command = new ArrayList<String>();
-      String actual = Utilities.path("[tmp]", "actual" + actualString.hashCode() + ".json");
-      String expected = Utilities.path("[tmp]", "expected" + expectedString.hashCode() + ".json");
-      TextFile.stringToFile(expectedString, expected);
-      TextFile.stringToFile(actualString, actual);
-      command.add(diff);
-      if (diff.toLowerCase().contains("meld"))
-        command.add("--newtab");
-      command.add(expected);
-      command.add(actual);
-
-      ProcessBuilder builder = new ProcessBuilder(command);
-      builder.directory(new CSFile(Utilities.path("[tmp]")));
-      builder.start();
-
-    }
-    return result;
-  }
-
-
-  private static String compareText(String expectedString, String actualString) {
-    for (int i = 0; i < Integer.min(expectedString.length(), actualString.length()); i++) {
-      if (expectedString.charAt(i) != actualString.charAt(i))
-        return createNotEqualMessage("Strings differ at character " + Integer.toString(i), String.valueOf(expectedString.charAt(i)), String.valueOf(actualString.charAt(i)));
-    }
-    if (expectedString.length() != actualString.length())
-      return createNotEqualMessage("Strings differ in length but match to the end of the shortest.", Integer.toString(expectedString.length()), Integer.toString(actualString.length()));
-    return null;
-  }
-
-  public static String tempFile(String folder, String name) throws IOException {
-    String tmp = tempFolder(folder);
-    return Utilities.path(tmp, name);
-  }
-
-  public static String tempFolder(String name) throws IOException {
-    File tmp = new File("C:\\temp");
-    if (tmp.exists() && tmp.isDirectory()) {
-      String path = Utilities.path("C:\\temp", name);
-      Utilities.createDirectory(path);
-      return path;
-    } else if (ToolGlobalSettings.hasTempPath()) {
-      return ToolGlobalSettings.getTempPath();
-    } else if (new File("/tmp").exists()) {
-      String path = Utilities.path("/tmp", name);
-      Utilities.createDirectory(path);
-      return path;
-    } else {
-      String path = Utilities.path(System.getProperty("java.io.tmpdir"), name);
-      Utilities.createDirectory(path);
-      return path;
-    }
-  }
 }
