+++ org.hl7.fhir.r5/src/test/resources/testUtilities/json/actualDiffArrayContent.json.error
@@ -1,1 +1,3 @@
-string property values differ at .expectedArray[0]: type expectedValue 1/unexpectedValue 1
+string property values differ at .expectedArray[0]
+Expected :expectedValue 1
+Actual  :unexpectedValue 1
