+++ org.hl7.fhir.r5/src/test/resources/testUtilities/json/actualDiffBoolean.json.error
@@ -1,1 +1,3 @@
-boolean property values differ at .expectedBoolean: type true/false
+boolean property values differ at .expectedBoolean
+Expected :true
+Actual  :false
