+++ org.hl7.fhir.r5/src/test/resources/testUtilities/json/actualDiffNumber.json.error
@@ -1,1 +1,3 @@
-number property values differ at .expectedNumber: type 123/789
+number property values differ at .expectedNumber
+Expected :123
+Actual  :789
