+++ org.hl7.fhir.r5/src/test/resources/testUtilities/json/actualDiffType.json.error
@@ -1,1 +1,3 @@
-property types differ at .expectedString: type expected value/1
+property types differ at .expectedString
+Expected :expected value
+Actual  :1
