+++ org.hl7.fhir.r5/src/test/resources/testUtilities/json/actualDiffValue.json.error
@@ -1,1 +1,3 @@
-string property values differ at .expectedString: type expected value/unexpected value
+string property values differ at .expectedString
+Expected :expected value
+Actual  :unexpected value
