+++ org.hl7.fhir.r5/src/test/resources/testUtilities/xml/actualDiffAttribute.xml.error
@@ -1,1 +1,3 @@
-Attributes differ at /root/blah: value dummyAtt/wrongwrongwrong
+Attributes differ at /root/blah
+Expected :dummyAtt
+Actual  :wrongwrongwrong
