+++ org.hl7.fhir.r5/src/test/resources/testUtilities/xml/actualDiffLocalName.xml.error
@@ -1,1 +1,3 @@
-Names differ at /root: nameSpacedNode/wrongNameSpacedNode
+Names differ at /root
+Expected :nameSpacedNode
+Actual  :wrongNameSpacedNode
