+++ org.hl7.fhir.r5/src/test/resources/testUtilities/xml/actualDiffNodeType.xml.error
@@ -1,1 +1,3 @@
-node type mismatch in children of /root/blah: 1/1
+node type mismatch in children of /root/blah
+Expected :1
+Actual  :1
