+++ org.hl7.fhir.r5/src/test/resources/testUtilities/xml/actualDiffText.xml.error
@@ -1,1 +1,3 @@
-Text differs at /root/blah: expected/different
+Text differs at /root/blah
+Expected :expected
+Actual  :different
