+++ org.hl7.fhir.r5/src/test/resources/testUtilities/xml/actualDiffTextEmpty.xml.error
@@ -1,1 +1,1 @@
-node mismatch - more nodes in target in children of /root/emptyNode
+node mismatch - more nodes in expected in children of /root/emptyNode
