+++ org.hl7.fhir.utilities/src/main/java/org/hl7/fhir/utilities/tests/BaseTestingUtilities.java
@@ -101,4 +101,28 @@ public class BaseTestingUtilities {
       }
     }
   }
+
+  public static String tempFile(String folder, String name) throws IOException {
+    String tmp = tempFolder(folder);
+    return Utilities.path(tmp, name);
+  }
+
+  public static String tempFolder(String name) throws IOException {
+    File tmp = new File("C:\\temp");
+    if (tmp.exists() && tmp.isDirectory()) {
+      String path = Utilities.path("C:\\temp", name);
+      Utilities.createDirectory(path);
+      return path;
+    } else if (ToolGlobalSettings.hasTempPath()) {
+      return ToolGlobalSettings.getTempPath();
+    } else if (new File("/tmp").exists()) {
+      String path = Utilities.path("/tmp", name);
+      Utilities.createDirectory(path);
+      return path;
+    } else {
+      String path = Utilities.path(System.getProperty("java.io.tmpdir"), name);
+      Utilities.createDirectory(path);
+      return path;
+    }
+  }
 }
